Improve torque, Update input, Minor cleaning
Physics is still broken. The problem was thought be be torque from the
normal impulse being applied in world space to the local space angular
velocity, but this appears to only be part of the whole problem. From
testing with methods of displaying the cubes, it appears that rotation
of objects is inexplicably processed differently in different pices of
code, notably collision detection and drawing will give very different
results with the same method. I expect this to be a much deeper problem
than physics but the fundimental problem of euler angles being ambiguous
in the order rotations should be applied in.

Input has been updated with a default constructor and some common input
values for movement, mouse position, etc.

+++ Shared/ClientObject.cpp
@@ -47,7 +47,7 @@ void ClientObject::updateStateWithInputBuffer(const PhysicsState& state, RakNet:
 
 	RakNet::Time lastTime = stateTime;
 	bool isFirstInput = true;
-	for (int i = 0; i < inputBuffer.getSize(); i++)
+	for (unsigned short i = 0; i < inputBuffer.getSize(); i++)
 	{ 
 		// The buffer contains a std::tuple containing time, physics state, and input
 		auto& input = inputBuffer[i];
